requirements.txt
Cython==0.25.2
glfw==1.3.3
imgui==0.0.0
multipledispatch==0.4.9
numpy==1.12.0
olefile==0.44
Pillow==4.0.0
psutil==5.2.0
pyassimp==3.3
PyOpenGL==3.1.0
PySDL2==0.9.5
recordclass==0.4.2
scanf==1.4
sip==4.19
